Renaming some stuff, writing some comments
changed board width / height and tile side length to fields on the
board, changed the functions that were being used to calculate these
values to getters.  These values are only set when a board is created,
so no need to calculate them each time.  If we want a new board, a new
object must be created.