Refactoring Level Start code out of Main.as
I took out all of the level starting code from Main.as and put it in a
function in Game.as, startLevel.  This function takes a string that will
specify the level to start.  Currently parseLevel (which is what this
string is passed to) just ignores it so oh well.